requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
pillow==11.1.0
psycopg2-binary==2.9.10
boto3==1.35.94
psycopg2-binary==2.9.10
channels==4.2.0
channels_redis==4.2.1
uvicorn[standard]==0.34.0
